refactor : Test DTO 패키지 분리 및 Lombok 적용
